# api.yaml
swagger: '2.0'
info:
  title: API микросервиса авторизации
  version: 1.0.0
paths:
...
  /auth/v1/changeLogin:
    post:
      tags:
        - Employees
      description: Запрос на изменение логина пользователя
      operationId: changeLogin
      consumes:
        - application/json
      parameters:
        - in: body
          name: body
          description: Данные для изменения логина пользователя
          required: true
          schema:
            $ref: '#/definitions/changeLoginData'
      responses:
        200:
          description: ОК
        403:
          description: Доступ запрещен
        500:
          description: Internal server error
...
  /auth/v1/private/setNewPassword/{userId}:
    post:
      tags:
        - Employees
      description: Запрос на установку пользователю пароля
      operationId: setPasswordWithOutPermission
      consumes:
        - application/json
      parameters:
        - in: path
          name: userId
          description: Идентификатор пользователя
          required: true
          type: integer
          format: int64
        - in: body
          name: setPasswordRq
          description: Данные для изменения пароля
          required: true
          schema:
            $ref: '#/definitions/setPasswordData'
      responses:
        200:
          description: OK
        400:
          description: Ошибка валидации
          schema:
            $ref: '#/definitions/errorData'
        404:
          description: Not Found
        500:
          description: Internal server error
  /auth/v1/private/changeLogin:
    post:
      tags:
        - Employees
      description: Запрос на изменение логина пользователя
      operationId: changeLoginWithOutPermission
      consumes:
        - application/json
      parameters:
        - in: body
          name: body
          description: Данные для изменения логина пользователя
          required: true
          schema:
            $ref: '#/definitions/privateChangeLoginData'
      responses:
        200:
          description: ОК
        403:
          description: Доступ запрещен
        500:
          description: Internal server error
  /auth/v1/employees:
    put:
      tags:
        - Employees
      description: Получение списка учетных записей всех сотрудников. Список отсортирован по ФИО сотрудников.
      operationId: getEmployees
      x-spring-pageable: true
      consumes:
        - application/json
      produces:
        - application/json
      parameters:
        - in: body
          name: filterData
          description: Тело запроса с данными фильтрации
          required: false
          schema:
            $ref: '#/definitions/filterEmployeesRq'
        - in: query
          type: boolean
          name: hasFilter
          description: Признак фильтрации сотрудников
          required: false
      responses:
        200:
          description: ОК
          schema:
            $ref: '#/definitions/employeeDataForList'
        403:
          description: Доступ запрещен
        500:
          description: Internal server error
    post:
      tags:
        - Employees
      description: Создание сотрудника
      operationId: createEmployee
      consumes:
        - application/json
      produces:
        - application/json
      parameters:
        - in: body
          name: createEmployeeRq
          description: данные сотрудника
          required: true
          schema:
            $ref: '#/definitions/createOrUpdateEmployeeRq'
      responses:
        200:
          description: ОК
          schema:
            $ref: '#/definitions/createEmployeeRs'
        400:
          description: Ошибка валидации
          schema:
            $ref: '#/definitions/errorData'
        403:
          description: Доступ запрещен
        500:
          description: Internal server error
  /auth/v1/private/employees:
    get:
      tags:
        - Employees
      description: Получение списка пользователей системы (private api)
      operationId: getEmployeesWithOutPermissions
      produces:
        - application/json
      parameters:
        - in: body
          name: filterData
          description: Тело запроса с данными фильтрации
          required: false
          schema:
            $ref: '#/definitions/filterEmployeesRq'
      responses:
        200:
          description: OK
          schema:
            type: array
            items:
              $ref: '#/definitions/employeeDataForList'
        500:
          description: Internal server error
  /auth/v1/private/employees/all:
    get:
      tags:
        - Employees
      description: Получение списка пользователей системы (private api)
      operationId: getEmployeesWithOutPermissionsWithDeleted
      produces:
        - application/json
      responses:
        200:
          description: OK
          schema:
            type: array
            items:
              $ref: '#/definitions/employeeDataForList'
        500:
          description: Internal server error
  /auth/v1/private/employee:
    get:
      tags:
        - Employees
      description: Получить пользователя системы по логину (private api)
      operationId: getEmployeeByLoginWithOutPermissions
      produces:
        - application/json
      parameters:
        - in: query
          name: login
          type: string
          required: true
          description: Логин пользователя
      responses:
        200:
          description: OK
          schema:
            $ref: '#/definitions/employeeDataForList'
        404:
          description: Not Found
        500:
          description: Internal server error

    post:
      tags:
        - Employees
      description: Создание сотрудника (private api)
      operationId: createEmployeeWithOutPermissions
      consumes:
        - application/json
      produces:
        - application/json
      parameters:
        - in: body
          name: createEmployeeRq
          description: данные сотрудника
          required: true
          schema:
            $ref: '#/definitions/createOrUpdateEmployeeRq'
      responses:
        200:
          description: ОК
          schema:
            $ref: '#/definitions/createEmployeeRs'
        400:
          description: Ошибка валидации
          schema:
            $ref: '#/definitions/errorData'
        403:
          description: Доступ запрещен
        500:
          description: Internal server error
  /auth/v1/employees/{id}:
    get:
      tags:
        - Employees
      description: Запрос получения информации о сотруднике
      operationId: getEmployeeById
      produces:
        - application/json
      parameters:
        - in: path
          name: id
          type: integer
          format: int64
          required: true
          description: Идентификатор сотрудника
      responses:
        200:
          description: ОК
          schema:
            $ref: '#/definitions/getEmployeeRs'
        403:
          description: Доступ запрещен
        500:
          description: Internal server error
    put:
      tags:
        - Employees
      description: Запрос обновления данных сотрудника
      operationId: updateEmployee
      produces:
        - application/json
      parameters:
        - in: path
          name: id
          type: integer
          format: int64
          required: true
          description: Идентификатор сотрудника
        - in: body
          name: updateEmployeeRq
          description: Тело запроса обновления данных пользователя
          required: true
          schema:
            $ref: '#/definitions/createOrUpdateEmployeeRq'
      responses:
        200:
          description: ОК
        400:
          description: Ошибка валидации
          schema:
            $ref: '#/definitions/errorData'
        403:
          description: Доступ запрещен
        404:
          description: Сотрудник не найден
        500:
          description: Internal server error
    delete:
      tags:
        - Employees
      description: Запрос удаления сотрудника
      operationId: deleteEmployee
      produces:
        - application/json
      parameters:
        - in: path
          name: id
          type: integer
          format: int64
          required: true
          description: Идентификатор сотрудника
      responses:
        200:
          description: Сотрудник удален
        404:
          description: Сотрудник не существует
        500:
          description: Internal server error
  /auth/v1/private/employees/{id}:
    get:
      tags:
      - Employees
      description: Запрос получения информации о сотруднике
      operationId: getEmployeeByIdWithOutPermission
      produces:
      - application/json
      parameters:
      - in: path
        name: id
        type: integer
        format: int64
        required: true
        description: Идентификатор сотрудника
      responses:
        200:
          description: ОК
          schema:
            $ref: '#/definitions/emploeeDataWithOrgUnits'
        400:
          description: Not Found
        500:
          description: Internal server error
    put:
      tags:
        - Employees
      description: Запрос обновления данных сотрудника (private api)
      operationId: updateEmployeeWithOutPermissions
      produces:
        - application/json
      parameters:
        - in: path
          name: id
          type: integer
          format: int64
          required: true
          description: Идентификатор сотрудника
        - in: body
          name: updateEmployeeRq
          description: Тело запроса обновления данных пользователя
          required: true
          schema:
            $ref: '#/definitions/createOrUpdateEmployeeRq'
      responses:
        200:
          description: ОК
        400:
          description: Ошибка валидации
          schema:
            $ref: '#/definitions/errorData'
        403:
          description: Доступ запрещен
        404:
          description: Сотрудник не найден
        500:
          description: Internal server error
  /auth/v1/private/user/{id}/update:
    post:
      tags:
        - Employees
      description: Запрос обновления основных полей пользователя (private api)
      operationId: updateUserWithOutPermissions
      consumes:
        - application/json
      parameters:
        - in: path
          name: id
          type: integer
          format: int64
          required: true
          description: Идентификатор пользователя
        - in: body
          name: updateUserRq
          description: Тело запроса обновления данных пользователя
          required: true
          schema:
            $ref: '#/definitions/updateUserRq'
      responses:
        200:
          description: ОК
        400:
          description: Ошибка валидации
          schema:
            $ref: '#/definitions/errorData'
        403:
          description: Доступ запрещен
        404:
          description: Сотрудник не найден
        500:
          description: Internal server error
  /auth/v1/employees/{id}/lock:
    post:
      tags:
        - Employees
      description: Запрос блокировки пользователя
      operationId: lockEmployee
      produces:
        - application/json
      parameters:
        - in: path
          name: id
          type: integer
          format: int64
          required: true
          description: Идентификатор сотрудника
        - in: body
          name: lockEmployeeRq
          description: Тело запроса блокировки пользователя
          required: false
          schema:
            $ref: '#/definitions/lockEmployeeRq'
      responses:
        200:
          description: ОК
        400:
          description: Ошибка валидации
          schema:
            $ref: '#/definitions/errorData'
        403:
          description: Доступ запрещен
        404:
          description: Сотрудник не найден
        500:
          description: Internal server error
  /auth/v1/employees/{id}/unlock:
    post:
      tags:
        - Employees
      description: Запрос разблокировки пользователя
      operationId: unlockEmployee
      produces:
        - application/json
      parameters:
        - in: path
          name: id
          type: integer
          format: int64
          required: true
          description: Идентификатор сотрудника
      responses:
        200:
          description: ОК
        400:
          description: Ошибка валидации
          schema:
            $ref: '#/definitions/errorData'
        403:
          description: Доступ запрещен
        404:
          description: Сотрудник не найден
        500:
          description: Internal server error
  /auth/v1/employees/import/sync:
    get:
      tags:
        - Employees
      description: Синхронизация списка учетных записей сотрудников с файлом
      operationId: syncEmployeeList
      produces:
        - application/json
      responses:
        200:
          description: ОК
          schema:
            $ref: '#/definitions/updateListStatus'
        403:
          description: Доступ запрещен
        500:
          description: Internal server error
  /auth/v1/employees/import/sync/status:
    get:
      tags:
        - Employees
      description: Получить статус синхронизация списка учетных записей сотрудников с файлом
      operationId: statusSyncEmployeeList
      produces:
        - application/json
      responses:
        200:
          description: ОК
          schema:
            $ref: '#/definitions/updateListStatus'
        403:
          description: Доступ запрещен
        500:
          description: Internal server error
  /auth/v1/changePassword:
    post:
      tags:
        - Employees
      description: Запрос на изменение пароля пользователя
      operationId: changePassword
      consumes:
        - application/json
      parameters:
        - in: body
          name: changePasswordRequest
          description: Данные для изменения пароля пользователя
          required: true
          schema:
            $ref: '#/definitions/changePasswordData'
      responses:
        200:
          description: ОК
        403:
          description: Доступ запрещен
        500:
          description: Internal server error
  /auth/v1/private/changePassword:
    post:
      tags:
        - Employees
      description: Запрос на изменение пароля пользователя
      operationId: changePasswordPrivate
      consumes:
        - application/json
      parameters:
        - in: body
          name: changePasswordRequest
          description: Данные для изменения пароля пользователя
          required: true
          schema:
            $ref: '#/definitions/changePasswordData'
      responses:
        200:
          description: ОК
          schema:
            $ref: '#/definitions/changePasswordRs'
        403:
          description: Доступ запрещен
        500:
          description: Internal server error
  /auth/v1/resetPassword/{id}:
    get:
      tags:
        - Employees
      description: Запрос на сброс пароля пользователя
      operationId: resetPassword
      parameters:
        - in: path
          name: id
          type: integer
          format: int64
          required: true
          description: Идентификатор сотрудника.

      responses:
        200:
          description: ОК
        403:
          description: Доступ запрещен
        500:
          description: Internal server error
  /auth/v1/private/client/resetPassword/{id}:
    get:
      tags:
        - Employees
      description: Установка технического пароля для клиента
      operationId: resetClientPasswordWithOutPermission
      parameters:
        - in: path
          name: id
          type: integer
          format: int64
          required: true
          description: Идентификатор клиента.
        - in: query
          name: password
          type: string
          required: true
          description: Технический пароль
      responses:
        200:
          description: ОК
        404:
          description: Not found
        500:
          description: Internal server error
...
definitions:
  changePasswordRs:
    type: object
    allOf:
    - $ref: '#/definitions/errorData'
...
  privateChangeLoginData:
    type: object
    required:
      - oldLogin
      - newLogin
    properties:
      oldLogin:
        type: string
        description: Текущий логин пользователя
      newLogin:
        type: string
        description: Желаемый логин пользователя
...
  errorData:
    type: object
    properties:
      errors:
        type: array
        items:
          type: object
          properties:
            fieldName:
              type: string
              description: Поле, к которому относится ошибка
            errorMessage:
              type: string
              description: Сообщение об ошибке
...
  getEmployeeRs:
    allOf:
      - $ref: '#/definitions/employeeData'
      - type: object
        properties:
          id:
            type: integer
            format: int64
            description: Идентификатор учётной записи сотрудника
          changePassword:
            type: boolean
            description: Признак необходимости сменить пароль при входе
          segment:
            type: string
            description: Название сегмента
          roles:
            type: array
            description: Назначенные роли
            items:
              type: integer
              format: int64
              description: Идентификатор роли
          lockStatus:
            type: boolean
            description: Признак, блокировки пользователя в системе
  employeeDataForList:
    allOf:
      - $ref: '#/definitions/employeeData'
      - type: object
        properties:
          id:
            type: integer
            format: int64
            description: Идентификатор учётной записи сотрудника
          fullName:
            type: string
            description: ФИО сотрудника
          branches:
            type: array
            description: Список наименований региональных филиалов
            items:
              type: string
              description: Название регионального филиала
          offices:
            type: array
            description: Список наименований офисов
            items:
              type: string
              description: Название офиса
          segment:
            type: string
            description: Название сегмента
          roles:
            type: array
            description: Назначенные роли
            items:
              type: string
              description: Имя роли
          lockStatus:
            type: boolean
            description: Признак, блокировки пользователя в системе
...
  filterEmployeesRq:
    type: object
    properties:
      secondName:
        type: string
        description: Фамилия сотрудника
      firstName:
        type: string
        description: Имя сотрудника
      middleName:
        type: string
        description: Отчество сотрудника
      birthDate:
        type: string
        format: date
        description: Дата рождения
      personnelNumber:
        type: string
        description: Табельный номер
      mobilePhone:
        type: string
        description: Номер мобильного телефона
      branches:
        description: Название регионального филиала
        type: array
        items:
          type: string
      orgUnitIdes:
        description: Название регионального филиала
        type: array
        items:
          type: integer
          format: int64
      position:
        description: Должность сотрудника
        type: string
      roleId:
        description: Роль сотрудника
        type: integer
        format: int64
      motivationCorrectStatus:
        $ref: '#/definitions/motivationCorrectStatus'
        description: Статус корректности мотивации
  changePasswordData:
    type: object
    required:
      - login
      - oldPassword
      - newPassword
    properties:
      login:
        type: string
        description: логин
      oldPassword:
        type: string
        description: старый пароль
      newPassword:
        type: string
        description: новый пароль
...
  createOrUpdateEmployeeRq:
    allOf:
      - $ref: '#/definitions/employeeData'
      - type: object
        description: Данные сотрудника
        required:
          - roles
          - firstName
          - secondName
          - personnelNumber
          - orgUnitId
          - segmentId
        properties:
          password:
            type: string
            description: пароль
          roles:
            type: array
            description: Назначенные роли (идентификаторы)
            items:
              type: integer
              format: int64
  createEmployeeRs:
    type: object
    properties:
      id:
        type: integer
        format: int64
        description: Идентификатор пользователя
...
  lockEmployeeRq:
    type: object
    properties:
      description:
        type: string
        description: Описание причины блокировки
  updateListStatus:
    type: object
    properties:
      status:
        type: string
...
  changeLoginData:
    type: object
    required:
      - login
    properties:
      login:
        type: string
        description: логин
  setPasswordData:
    type: object
    required:
    - newPassword
    - expireTime
    properties:
      newPassword:
        type: string
        description: Пароль
...